Create Board_State_Class.cpp
The implementation of the Board_State Class. This class is used to represent the state of the chess board. Games can be loaded into the engine by passing the fen notation in as an argument, in the form of a string. If no game is loaded in, the starting position is generated.